+++ CHANGELOG.md
@@ -5,6 +5,7 @@ HEAD
 ----
 
 * Fix configuring Mqtt debug message
+* Fix return for getTopicSub method when the topic was not found
 
 v0.0.8 (2023-05-09)
 ------