+++ CHANGELOG.md
@@ -6,6 +6,7 @@ HEAD
 
 * Fix configuring Mqtt debug message
 * Fix return for getTopicSub method when the topic was not found
+* Change mqtt_client_event_data.data_len to size_t
 
 v0.0.8 (2023-05-09)
 ------