+++ CHANGELOG.md
@@ -11,6 +11,7 @@ HEAD
 * Remove conflicting defines for esp8266
 * Remove WebConfigServer from basic_class_with_mqtt and update example with latest version
 * Add basic_class_with_mqtt_WebConfigServer example
+* Exclude .github folder for pio pkg
 
 v0.0.8 (2023-05-09)
 ------